Task 13. Delete states

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -16,5 +16,6 @@ if __name__ == "__main__":
 
     my_session = Session()
     my_query = my_session.query(State).filter(State.name.like('%a%'))
-    my_query.delete()
+    if my_query.count >=1:
+        my_query.delete()
     my_session.commit()
